extract server code into separate directory
