+++ src/shared/lib/hashing-utils.ts
